+++ Addon/ElvUI_mMediaTag/options/options_core.lua
@@ -586,15 +586,8 @@ mMT.options = {
 					childGroups = "tab",
 					args = {},
 				},
-				classcodex = {
-					order = 4,
-					type = "group",
-					name = mMT:AddSettingsIcon("Class Codex", "book"),
-					childGroups = "tab",
-					args = {},
-				},
 				premade_groups_filter = {
-					order = 5,
+					order = 4,
 					type = "group",
 					name = mMT:AddSettingsIcon("Premade Groups Filter", "dungeon"),
 					childGroups = "tab",